+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 
 ### Added
 - quickCheck Parser test for arguments. Arbitrary instances for Language.GraphQL.AST.Document.
+- Enhanced query error messages. Add tests for these cases.
 
 ## [1.0.2.0] - 2021-12-26
 ### Added